Basic metrics, benchmarks
This commit removes the tracing integration and uses only metrics.
We were only really using metrics and figure this will be how we
emit for long-term storage. Prometheus-only for now.

Benchmark for fill_ascii_buffer is in place. Will shortly experiment
with faster RNG algorithms. Current 600Mb/s buffer filling is pretty
poor for this program.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>