Generate json
This commit allows for the generation of json. It's not especially
fast, 1/3 the speed of just text generation, but no better approach
has crossed my mind. It's... not quite right. The lines in the json file
are constant-ish sized where the user expects to be able to manipulate the
size of lines. Currently they manipulate the batch size of lines.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>