Correct "bytes_written" telemetry (#6)
Internal to file_gen's LogTarget is the notion of bytes_written, which is what
tracks progress toward the maximum file size of the target. Also there is a
bytes_written that is output as telemetry which really tracks the bytes filled
into the buffer, not the global size. Unfortunately because of name confusion I
got the whole mess confused and it looked like, from telemetry, that byte output
rate was getting faster and faster. Provably not accurate by looking at the
file.

Closes #4

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>