Pivot to the use of Arbitrary (#33)
This commit contains some interesting work. Firstly, I have replaced the
`buffer` code with the use of `Arbitrary`, reducing the project's dependence on
wonky custom generation code. I have also removed any `BufWriter` in the spin
loop, pivoting instead to the use of fixed MiB sized "blocks" that are
constructed ahead of time before entering the spin loop. This slows down
construction some -- hence the introduction of rayon -- but does significantly
improve on file_gen's throughput capability.

The config file has changed somewhat. We'll need to address the loss of line
length control but this work feels overlong as is.

Resolves #32
Resolves #31
REF #10

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>