Fix malformed block creation
Previously it was possible for file_gen to crash when generating new blocks as a
result of those blocks outpacing the governor. To resolve this I've hidden the
Arbitrary definition behind a slightly reworked Serializer trait. The caller is
now responsible for setting up instances of Serializer and nothing more.

Startup is now slower, unfortunately, because the Serializer isn't Sync + Send
necessarily, but that couldn't be helped.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>