Introduce `http_blackhole` (#59)
This commit introduces a new program, `http_blackhole`. It is what it sounds
like, a spot for HTTP traffic to sink into. It only ever responds 200 OK to all
traffic with a blank body when that's relevant. I have moved all the blackhole
programs into a single crate and have a notion we might do the same with
generation.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>