Allow user to set block sizes (#80)
* Allow user to set block sizes

This commit allows the user to set block sizes for generators. This does not
allow them to adjust individual message sizes -- unless a block is equivalent to
a message -- but does allow indirect control. Essentially, the larger the block
the larger the message for some formats.

This is mostly useful for systems that have a hard cap on the total payload size
-- say, UDP when we get to that -- or in situations where we need to simulate
consistent producers a little more carefully.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* clippy ding

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Allow block sizes to be defaulted

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* clippy dings

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>