Convert lading configuration from toml to yaml, remove rayon (#119)
Convert lading configuration from toml to yaml, remove rayon

This commit makes a breaking change to lading and opens up the 0.6 release
series. The major change here is the conversion to the use of yaml in lading's
configuration file, driven by the toml crates inability to deal with
deserializing enums that have values in them. This means our 'static' format
variant is not actually usable in toml-land, as that variant embeds a path in
it.

I have also taken the opportunity to do a little house cleaning, removing the
rayon dependency from file_gen and adjusting the goofy case issues that we've
had in configuration for a minute. Now every option will be snake_case and users
don't have to deal with the odd Post etc anymore.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>