Make randomness require an explict seed
While working on https://github.com/vectordotdev/vector/pull/10730/ it became
clear that one potential source of noise was lading itself. Consider if in one
run we randomly generate a payload that is unusually slow for a certain vector
configuration but never generate that same payload otherwise. This commit
requires that the user set explicit seeds for any randomness that lading uses.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>