Combine the sub-crates (#148)
This commit pretty substantially reworks the project. Over time we've accumulated a number of small binaries which, in [vector](https://github.com/vectordev/vector) we link up together in a minikube environment to do performance testing. That has worked but has a complexity burden. It occurred to me that if lading could run binaries as sub-processes a good deal of that burden could be removed, and we could do further interesting things like `perf` the sub-process to make flamegraphs etc etc. 

I set out to make a new `rig` binary in this project -- whence the branch name -- do that kind of thing and in the doing of it realized I could combine all the sub-crates and get a cleaner, more focused lading.

The changes here have been validated in https://github.com/vectordotdev/vector/pull/11891.