Rework the lading UX
This commit reworks the lading UX so that all the target details can be set from
the command line, and must be. This addresses an awkward situation where _part_
of the target could be configured from CLI but part still had to be present in
the configuration file. This supports workloads that do not have stable
locations for files.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>