Rework the lading UX (#150)
* Rework the lading UX

This commit reworks the lading UX so that all the target details can be set from
the command line, and must be. This addresses an awkward situation where _part_
of the target could be configured from CLI but part still had to be present in
the configuration file. This supports workloads that do not have stable
locations for files.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Update the UX, jump to 0.9.

This commit updates the project UX to use clap instead of argh and makes a few
serious changes. They are:

 * The target can now only be configured via CLI flags, giving more flexibility
   to the location of target paths etc and making it easier to know what flags are
   or are not set for the target.
 * I have added duration flags for the experiment and also warmup periods.
 * The 'index' line is renamed 'fetch_index' to match a predecessor of lading,
   avoid complications with data analysis tools that inject their own 'index'
   columns.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>