Collect perf report data from target process
This commit sets up a perf record sub-process that monitors the target process
if the user gives the `--target-perf-data-path` flag. I haven't done anything to
ensure this will actually work, which it won't depending on your Linux sysctl
settings and will never work on OS X or Windows. Still. I have had to undo some
of #156 so that perf can shutdown correctly, pull in Nix to avoid SIGKILL
signals to both the target and perf. This means the PR is blocked until #155 is
resolved.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>