Convert from quickcheck to proptest, add lints
This commit converts from the quickcheck library to proptest. I've
made a note of situations in which the block creation routine will
kick out blank blocks, which remains unresolved. I have also added
further lints to the project. `missing_docs` I have not flagged on yet
as there are lots of missing docs but that'll come.

Test coverage remains low, befitting a project born of a quick
prototype.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>