Collect CPU, memory, scheduling information about the target process
This commit introduces procfs dependency to the project and uses it to power an
'observer' server that runs as a peer to the target etc. This observer is
responsible for periodically reading the /proc/PID/stat of the target process,
writing out interesting details regarding memory consumption, CPU and scheduling
time. In this commit we collect:

* page_size -- the memory page size of the OS in bytes
* ticks_per_second -- the OS ticks that happen in a second
* rss_pages -- the real memory pages the process has allocated
* rsslim_bytes -- the soft limit of RSS in bytes
* num_threads -- the number of threads the process has running
* utime_ticks -- the number of ticks the process has spent scheduled in user-space
* stime_ticks -- the number of ticks the process has spent scheduled in kernel-space
* vsize_bytes -- the size in bytes of the process' virtual memory

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>