Resolve SplunkHec generator freeze bug
In https://github.com/vectordotdev/vector/pull/12776 we've seen that
the splunkhec generator doesn't quite work right with regard to
shutdown. This has been a long-running problem, see #155, and this
commit is intended to refactor the generator to be a little more
direct but, more importantly, have timeouts and definite loop breaks
in key places.

Whether this actually functions as expected will depend on Vector
where reproduction is the most feasible.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>