Allow static payload to take a directory (#221)
* Allow static payload to take a directory

This commit modifies the static generator to take either a file or a
directory. The block sizes are used as filters now, meaning we no longer assume
files can be arbitrarily truncated on line boundaries and instead pass the whole
file forward if `max_bytes` is sufficient to do so. It is up to the preparer of
static data to guarantee sufficient entropy.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* clippy dings

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* info debugging

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* fixes

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* fixes

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>