Allow file_gen users to decide on rotation
Previously the file generator would write to the end of its file, delete the
file and then start over with the same name. This caused some tailing programs
to perform strangely. That said, it's not necessarily interesting to _always_
remove handles and in some cases users would like to experiment with the tailer
doing file removal.

This is now possible. File generator will now always increment the file index
whether rotation is on or not but does allow the user to decide if rotation
happens, which it does by default.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>