Prepare lading for move under Datadog org.
I originally started lading as a side project, a little hack to write up for
technical writing purposes but found that it proved useful in my work on the
Vector project. Over time lading has got more and more useful to me and it's
more and more awkward to have it under my personal user.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>