Unify Payload configuration (#230)
Back when this project first started there were not many different kinds of
payload and only two generators. There was an awkward config -> payload switch
table that happened in each generator and, while I did love it, eh. There are
now more generators and more payloads with more on the way and what was once an
annoyance will quickly become a real code duplication problem.

This commit follows the generator/blackhole configuration model and has payload
provide its own configuration. This allows all the generators to drop their
Variant configuration and means introducing a new payload is not a big slog. If
we need generators to serialize differently from one another that is now more
difficult, but is surely something we can address by fiddling with the
Serializer trait.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>