Attempt to recreate UDS stream if not available, rather than crash
This commit adjusts the UDS generator spin implementation to be more like the
UDP implementation in that we no longer crash a subtask if the UDS socket is not
available for opening yet. In instances where Datadog Agent is concerned the
socket can be available relatively late, harming reliable inspection.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>