Ensure splunk_hec spawned tasks obey shutdown (#449)
This commit is peer to #447, #448. I have not had to thread a `select!` through
here as in the other commits but I did realize, following on from the file
generator work, that we could make shutdown of sub-tasks a little more
orderly. Considering that the splunk hec generator hangs lading shutdown
sometimes this is a welcome adjustment.

REF SMP-329

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>