Prefer to peek blocks
It's long nagged at me that we're too aggressive about popping blocks of a block
cache. Consider that previously we would pull a block off the cache, realize we
had no connection, drop the block, establish a connection and _then_ pop a block
and use that. It's not wasteful per se since we're cycling through no matter
what but it doesn't sit right, considering that the goal is for our cycle to be
consistent each time.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>