Allow for back-off based on target memory limits (#445)
* Allow for back-off based on target memory limits

This commit adds the ability for lading generators to back off from pushing load
if the target program gets close to a user-supplied memory limit. As of this
commit the work is not near complete, I have only fiddled with a handful of the
generators.

I think now it might be useful to make the block abstraction, well, abstract. If
the `BlockCache` allowed for peeking we wouldn't have to cycle blocks that never
get used, for instance. I think all the inner generator loops also need to move
to the `select!` model, so there will be some follow-up PRs before this one is
ready.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* add backoff, more metrics

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* clippy fixes

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

---------

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>