Unify limiting under a `Throttle` type
This commit takes our use of governor crate as well as the recent RSS limiter
and hides it inside a `Throttle` type that all generators now reference. I have
needed to spread `thiserror::Error` implementations around through this work,
which also unifies that aspect of our error handling.

There should be no happy-path functional change here. The ambition is to allow
us to hunt for setpoints in generation but making changes in ten distinct
generators to support that goal is, well, a dead-end.

REF SMP-372

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>