Remove `assert_fs` dependency
In #491 we ran into a dependency update issue revolving around `assert_fs` in
sheepdog, as well as that crate using a `tempfile` version dinged by
RUSTSEC-2023-0018. We have issued https://github.com/assert-rs/assert_fs/pull/93
to that crate but it also is the case that we can use `tempfile` directly, it
seems, here and remove the dependency on `assert_fs`.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>