Introduce support for TraceAgent payload
This commit introduces support for the Datadog Agent tracing span, defined
[here](https://github.com/DataDog/datadog-agent/blob/main/pkg/trace/pb/span.proto). In
practice this will be combined with the HTTP generator to POST payloads toward
the Agent.

As of this commit -- as noted in the code -- the shrink algorithm in our payload
creation isn't quite right, but I'm sure something will come to me tomorrow.

REF SMP-410

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>