Convert `Block` to use `Bytes` type not `Vec<u8>`
The `Bytes` type is really useful when you need to share the underlying
`Vec<u8>` without modification and our `Block` type is basically a fancy holder
for `Vec<u8>` with continual clones. The idea for this occurred to me in #506
although I didn't want to include it there.

There might be other locations in our codebase where `Bytes` is more appropriate
than `Vec<u8>`.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>