Allow for customizable payloads
This commit starts converting our payload generation away from the `arbitrary`
crate. While it's served us well for a long time it lacks in a few areas:

* customization by the user of data in-payload,
* ability for generation to reference 'up the tree'

These two points are especially useful for payloads like dogstatsd where we
would want to specify the number of metric names and the total number of unique
tag combinations.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>