Improve trace-agent payload generation speed
This commit determines by experimentation that our search for a payload size of
the just right variety was too stingy: we could be much more agressive. This is
done. Upfront we compute a tree of 10k members and then each time we find that
the tree is too small we tack on an additional 5k.

Extracted from #539
REF SMP-463

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>