Slightly tweak how observer performs math operations
This commit slightly tweaks how we tally up our tick and number of thread
data. Instead of immediately converting to a u64 we sum every interior
i64 and only then take the unsigned absolute value. The code is a little
more tidy than it was and at the extreme end of the value range we avoid
losing some information.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>