Update observer to collect CPU utilization data
This commit updates the observer with the aim of calculating CPU
utilization. We maintain output of user and kernel space CPU seconds
but add utilization with the understanding that this is useful to users
that need to consider utilization and not ticks.

Note, if there is CPU use in the system that is not in a parent, child
relationship but should still be counted this change will not address
that.

REF SMP-612

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>