Allow users to configure metric multivalue weights
In DogStatsD it is allowable for a metric line to have multiple values.
In practice, this is rare. Our previous 1/32 odds of producing only
a single-valued line makes lading's output unusual. This is not a bad
thing but it is behavior that we want users to be able to configure.

For the purposes of backward compatability we maintain the same 1/32
odds.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>