Introduce property tests for stable throttle
We have, lately, wondered if the stable throttle was giving too much capacity to
its clients. This PR is an attempt to answer that question through property
testing. Note I have separated out the async and non-async bits of `Stable` and
am property testing the non-async bit, making it feasible to do this kind of
examination. I suspect I can do the same for the predictive throttle.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>