Shift throttle code to lading_throttle, kani
I would really like to establish that the stable throttle behaves the way that
we expect. This commit does a few things. Firstly we move the throttle code to a
crate with minimal dependencies, allowing us to run the kani proof tool on that
new crate. Second, I introduce a kani experiment for the major property
test. This takes... some time to run and as of this writing I'm not sure of the
results. Finally, the predictive throttle is gone. In practice it wasn't used
because its results were too surprising.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>