Introduce property tests for stable throttle (#660)
* Introduce property tests for stable throttle

We have, lately, wondered if the stable throttle was giving too much capacity to
its clients. This PR is an attempt to answer that question through property
testing. Note I have separated out the async and non-async bits of `Stable` and
am property testing the non-async bit, making it feasible to do this kind of
examination. I suspect I can do the same for the predictive throttle.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* add test config

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* always check granted requests

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* Shift throttle code to lading_throttle, kani

I would really like to establish that the stable throttle behaves the way that
we expect. This commit does a few things. Firstly we move the throttle code to a
crate with minimal dependencies, allowing us to run the kani proof tool on that
new crate. Second, I introduce a kani experiment for the major property
test. This takes... some time to run and as of this writing I'm not sure of the
results. Finally, the predictive throttle is gone. In practice it wasn't used
because its results were too surprising.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* PR review

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* CI fixes

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* property ding

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* checkpoint

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* checkpoint

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* pass static

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* checkpoint

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* further shrink

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* checkpoint

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* draw down on the interval cell, not add to

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* re-expand experiments, fix kani proof

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* 0.17.5-rc1

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* comment fix

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* update README

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* 0.18.0

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* install protobuf in release

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

---------

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>