Satisfy clippy, make a `lading_rev` optimization target
This commit satisfies clippy in a handful of areas but most importantly adds a
`lading_rev` binary to the project. This binary will only be compiled when the
`dogstatsd_perf` feature flag is added to the build and is meant to be used for
optimizing the member generator.

Of note, we've started to discuss streaming built `Member` instances direclty to
a generator without going through a block cache. To this point we've assumed
that the member generation only needs to be "fast enough", hence all the cloning
and small string allocations and the like. The generator is _slow_ as a
result. Consider that if you compile the project now with

```
> cargo build --release --features dogstatsd_perf --bin lading_rev
...
> hyperfine
Benchmark 1: ./target/release/lading_rev
  Time (mean ± σ):     32.688 s ±  3.813 s    [User: 28.507 s, System: 2.650 s]
  Range (min … max):   30.190 s … 40.988 s    10 runs
```

Call it roughly 30ms per `Member` instance, an eternity. I'm working on a Mac so
running

```
> cargo instruments --release --features dogstatsd_perf --bin lading_rev -t time
```

does appear to show that we spend _a lot_ of time making and cloning small
strings, which is accurate to my understanding of the code as it exists
today. `AsciiString` is a particular culprit. 67% of program runtime is in
`impl payload::Generator<Tagsets>` as of this commit.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>