Shave some runtime
This commit removes the use of `choose_multiple` in the generator
`AsciiString`. Profiling shows that in a run of `lading_rev` this generator is
responsible for ~70% of program runtime. In _that_ ~60% of runtime is
`choose_multiple`. While convenient, it does turn out that the function
implicitly allocates a little `Vec` which we do not really need here, since
we're immediately pushing into storage otherwise.

```
➜  lading git:(sopell/dogstatsd-generator-revamp) ✗ hyperfine --warmup 3 ./target/release/lading_rev
Benchmark 1: ./target/release/lading_rev
  Time (mean ± σ):     20.902 s ±  0.167 s    [User: 18.438 s, System: 2.405 s]
  Range (min … max):   20.727 s … 21.279 s    10 runs
```

Shaves ~10 seconds off.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>