Move `lading::payload` into its own crate
This commit is inspired by the work done in #666. We would have a much easier
time writing benchmarks for payload generation _if_ the payloads were isolated
to their own crate. Note that in #666 I introduced a `lading_rev` binary to work
around this somewhat but the result is not totally satisfactory.

REF SMP-664

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>