Introduce benchmarks for `lading_payload`
In #671 we moved the payload module into a new crate in order to make it easier
to add benchmarks for payload generation. This is now done, we have benchmarks
for dogstatsd. Of note, we benchmark setup time -- ~30ms -- and throughput, the
later shows the real influence of our longish setup time: the more bytes we
generate the better the throughput.

It does strike me that the `Serialize` trait will not be suitable for
streaming, REF #666. Some payloads like dogstatsd can stream per-message but
others require serialization into a container. How to deal with the container
notion in a stream, I'm not sure. Something to think about.

REF SMP-673

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>