Introduce a pool mechanism for producing small strings
This commit is born of the work on #666. We realize in that PR that we spend
most of our runtime producing small heap allocated strings. This PR introduces
`Pool` which allows the user to request a small `&str` of a given size. It is
hooked up to only a single payload -- Ascii -- but will be introduced elsewhere
in follow-up work.

REF SMP-664

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>