Convert DogStatsD payload to use string pool
This commit begins the process of converting the dogstatsd payload generation to
use the new string pool, introduced in #675. I have only partially converted the
event sub-payload -- note that tagsets have not been switched yet -- and this
shows a 6% improvement to `dogstatsd_setup` and a 5% - 74% improvement to
`dogstatsd_all`. It shows there's some promise to the technique that improves as
we scale up the total-bytes emitted.

Of note I will need to eventually convert the `Generator` trait to emit a type
with a lifetime. Since I can't do that incrementally the `generate` function in
select areas temporarily does not come from the trait. I'll resolve this as a
part of the work here.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>