Convert DogStatsD payload to use string pool (#676)
* Convert DogStatsD payload to use string pool

This commit begins the process of converting the dogstatsd payload generation to
use the new string pool, introduced in #675. I have only partially converted the
event sub-payload -- note that tagsets have not been switched yet -- and this
shows a 6% improvement to `dogstatsd_setup` and a 5% - 74% improvement to
`dogstatsd_all`. It shows there's some promise to the technique that improves as
we scale up the total-bytes emitted.

Of note I will need to eventually convert the `Generator` trait to emit a type
with a lifetime. Since I can't do that incrementally the `generate` function in
select areas temporarily does not come from the trait. I'll resolve this as a
part of the work here.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* update service-check

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* Add more string pool to dogstatsd

This commit improves the string pool spread in dogstatsd. I still have yet to
adjust the metric generation to avoid cloning -- I'll do that in the next commit
-- but setup is improved by 34% and _all from 52% to 5%. I suspect that if I can
get the generation side to note clone we'll improve on the high-end.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* separate template notion in dogstatsd metrics

This commit makes an explicit `Template` that is used to generate a full
`Metric<'a>`. Note the new lifetime, we avoid cloning so much now. At the top
end this hits 1Gb/s in the _all benchmark.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* Do more writing in lading_rev

After instrumentation it appears that lading_rev spends 85% of its time in
`fmt::write`, implying that if we want to go faster we'll need to make coercion
into strings cheaper. Excellent result.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

---------

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>