Place block construction in a `Cache` struct (#684)
This commit puts the block construction into a type, called `Cache`. We support
a peek interface which requires mutable access and an indexing implementation
which is read-only. This is not quite pleasing to the eye but it maintains the
interfaces that already exist. This work is intended to be the basis of an
infinite streaming `Cache`, not enabled by default. That work will be done in a
follow-up PR.

For the most part this change is mechanical. The unix-datagram generator now
follows the file-gen in sharing a single `Cache` instance per writer, rather
than having one for each child.

REF SMP-664

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>