Block cache with an infinite stream option (#687)
* Block cache with an infinite stream option

This commit makes two major changes to the block cache. The most important is
that the block cache's major CPU consumption -- the construction of `Block`
instances -- is now done in a separate OS thread from the tokio runtime. This
allows us to introduce the second more important change: infinite streams of
`Blocks`. It is now possible for users to construct an unending stream of
`Block` instances that do not loop. We maintain a cache of constructed `Block`s
up to the maximum total bytes allow to minimize any potential latency impact.

Configuration is changed but not in a backward incompatible way.

REF SMP-664

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* v0.18.1-rc0

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

---------

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>