Adjust proofs for stable valve in throttle (#721)
* Adjust proofs for stable valve in throttle

This commit adjusts the proofs that we have in our stable throttle to avoid a
loop and instead rely on a series of smaller proofs. This terminates in a
reasonable amount of time. Note I've removed the proptests and this commit will
need to be followed by another than runs kani in CI.

I may still be missing a few small properties.

REF SMPTNG-79

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* add CI integration

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

---------

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>