Migrate `block` into `lading_payload`
This commit is intended to allow for us to experiment more deeply with proof
tools, fuzzing and the lading block mechanism. I have migrated block from the
main lading crate into `lading_payloads` as kani currently cannot build the main
lading crate and, second, I want to expose the block as a public interface for
fuzzing.

REF SMPTNG-79

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>