Introduce proofs for block mechanism (#725)
* Introduce proofs for block mechanism

This commit adds proofs for the `chunk_bytes` function, converting its interface
to be in terms of `u32` -- as matches our configuration -- and also to use a
round-robin method internally to avoid randomness. I have also considered having
the caller responsible for passing `chunks` in as a mutable reference, allowing
us to bound the maximum allocation of this function, currently unbounded. That
would also make this function allocation-free, which would be a neat trick,
although we'd only be shuffling the allocation around, not removing it.

REF #724
REF SMPTNG-79

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* Adjust implementation to not allocate, proofs

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* Add additional proof, improve language

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

---------

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>