procfs: start new generator
We'd like to be able to use a generator that generates a `/proc`
filesystem in order to generate load for observability tools that
query that filesystem. This commit starts the process of writing a
`procfs` generator in order by adding some constants, an `Error` enum
with a variant I know will be used eventually, and some data
structures intended to model some files in `/proc` for processes.

Some of the data structures -- notably `Status` and `Process` -- are
incomplete, and require more fields and documentation. This commit is
an end-of-day commit to communicate progress so far.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>