procfs generator: add remaining fields of status
The previous end-of-day commit for the `procfs` generator had not
added all of the fields needed to generate `/proc/[pid]/status`; this
commit adds the remaining fields, and in addition, updates the types
of fields to model more accurately the underlying data in the Linux
kernel used to populate those fields.

On the data modeling side of things, there is still some work to do on
`/proc/[pid]/status` because kernel configuration settings determine
the presence or absence of some of the fields in `/proc/[pid]/status`.

In addition, I think `/proc/[pid]/stat` will have to be modeled with a
separate struct because it seems to include information that
`/proc/[pid]/status` lacks (and vice versa; neither file's information
is a subset of the other's, even after accounting for differences in
formatting and differences in data views).

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>