procfs generator: more notes, sketch fake pid
This commit adds more notes and some comments for questions I need to
ask tomorrow to try and get a handle on the last bits of information
needed to set up a fake process as part of implementing a procfs
generator.

In particular, this commit documents how to generate
`/proc/[pid]/statm` as a view of `/proc/[pid]/status`, and also
documents the loose coupling between `/proc/[pid]/cmdline` and
`/proc/[pid]/comm` (the latter of which is related to the "name"
fields of `/proc/[pid]/status` and `/proc/[pid]/stat`.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>