Signal shutdown, use runtime shutdown to stop
This commit removes the soft shutdown that we previously used in lading. We have
known for some time that lading's shut down mechanism is faulty: singular tasks
would hold the shutdown up for unknown reasons. We then used the runtime
shutdown in addition to our soft mechanism.

This commit removes the soft mechanism. We signal that all tasks _should_ shut
down but no longer wait on them to do so, immediately moving to the runtime's
shutdown mechanism.

REF SMPTNG-94

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>