procfs payload: add process generator
Ultimately, we want to add a generator that generates data modeling a
"process", which is supposed to represent data for the following five
files:

    * /proc/[pid]/cmdline

    * /proc/[pid]/comm

    * /proc/[pid]/io

    * /proc/[pid]/stat

    * /proc/[pid]/statm

    * /proc/[pid]/status

This commit adds such a generator.

Implementing this generator exposed ownership issues when using string
slices as members in the `Stat`, `Status`, and `Process`
types. Earlier design conversations suggested that using the `String`
type as a member might be undesirable because of the number of heap
allocations it incurs. To avoid using the `String` type, this commit
replaces the `&'a str` members in those types with fixed-length arrays
of `u8`s. Using fixed-length byte arrays is feasible because the
underlying data types used by the Linux kernel to store the
corresponding data are also essentially byte arrays (really,
fixed-length arrays of `std::ffi:c_char`). This change also means it's
possible to use the `#[derive(Copy)]` macro on the various types
modeling those five files in `/proc/[pid]`, which might be useful when
implementing serialization.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>