procfs payload: have Process use `String`s
Using fixed-size byte buffers to store strings in the various types we
use seems, well, error-prone. I realized that the `dogstatsd` payload
uses `String`s, although it uses them with string slices that have
length bounded above by some value. The same idiom can be used in the
implementation of the `ProcessGenerator`, so this commit uses that
idiom.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>