procfs payload: impl Dislay for `/proc/[pid]/io`
I'm still trying to figure out what I'd like to do regarding
serializing the various `procfs` files. For now, as a stepping stone
to a proof-of-concept, this commit implements `fmt::Display` for
`lading_payload::procfs::proc::Io` as a kludgy way of simply getting
the data into a string in a relatively composable way. I may need to
rewrite this work later, but I should be able to salvage most of the
work from this commit if I need to rewrite it.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>