procfs payload: impl Display for Statm
Continuing with the theme of "write functions that convert data types
into strings", this commit adds a `fmt::Display` implementation for
the `lading_payload::procfs::Statm` type that models
`/proc/[pid]/statm`. The basic idea is to hedge against the
uncertainty I have in the serialization implementation by writing
something that will stringify the data model I have in a composable
fashion.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>