procfs payload: add task state helper methods
We're going to need to serialize the task state information in
`lading_payload::procfs::proc::State` for `/proc/[pid]/stat` and
`/proc/[pid]/status`, but each of these files uses a slightly
different format. The former users a single-letter code, whereas the
latter uses a longer, more human-readable string that starts the the
single-letter code. This commit implements helper methods that output
the string representation of the task state for each format.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>