procfs payload: impl Display for Stat
Further continuing the theme of writing functions that output string
representations of files in `/proc/[pid]/`, this commit implements
`std::fmt::Display` for `lading_payload::procfs::Stat`, which emits a
string that would be the contents of `/proc/[pid]/stat`. The working
premise for using `std::fmt::Display` is that it could be cannibalized
easily if it needs to be refactored into something else.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>