procfs payload: add cpuset format helper types
The "list format" and "mask format" cpuset lists for the
`{Cpus,Mems}_allowed` and `{Cpus,Mems}_allowed_list` types need to be
stringified, and I'd rather not write bespoke functions to do that if
I could exploit the trait system to do the same thing in a way that
seems more composable. This commit adds helper types to serialize
those lists to strings via `std::fmt::Display` implementations.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>