procfs payload: add Display impl for Status
We're going to have to write some string representing the contents of
`/proc/[pid]/status` to that file in order to complete a `procfs`
generator. The `lading_payload::procfs::Status` type is supposed to
model the contents of that file, so implementing the
`std::fmt::Display` trait for that type should enable us to stringify
the type (via invoking the type's `to_string` method) for writing into
a file or for some other serialization task. This commit implements
that trait for that type.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>