Hook procfs up into the generator, allow for end-user use (#773)
* Hook procfs up into the generator, allow for end-user use

This commit completes the work begun in #740 and exposes the purely faked procfs
to end users. Nothing very clever is done in this PR. The only novelty is we
avoid the use of a block cache and write stringy versions of the generated
structs directly to disk. The 'server' constructed here idles entirely.

REF SMPTNG-88

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* fixes

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* .3-rc0

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* allow copy verbatim from host proc

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* 0.20.3

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

---------

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>