changelog: prep for 0.20.4 release (#776)
* changelog: prep for 0.20.4 release

Before cutting a `lading` release, its changelog must be updated. This
commit updates the changelog in preparation for a 0.20.4 release.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>

* changelog: actually use 0.20.4-rc1, not 0.20.4

Changelog needs to have release candidate version to pass CI, so this
commit rectifies that oversight.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>

* Cargo.toml: bump version to 0.20.4-rc1

The `lading` package version needs to be bumped in `Cargo.toml`, so
this commit does that.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>

* changelog for 0.20.4: add more PR changes

Many more significant changes have been made since version 0.20.3 was
released. This commit adds more of those changes to the changelog for
version 0.20.4, and removes the `-rc1` release candidate designation.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>

* lading/Cargo.toml: change version to 0.20.4.

In preparation for a production release, this commit drops the "-rc1"
from the version of `lading` in `lading/Cargo.toml`.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>

---------

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>